Refactor: 타임스탬프 유틸 함수 분리
